src/requirements.txt
amqp==2.5.1
billiard==3.6.1.0
Django==2.2.7
django-widget-tweaks==1.4.5
importlib-metadata==0.23
kombu==4.6.5
more-itertools==7.2.0
Pillow==6.2.1
pytz==2019.3
redis==3.3.11
sqlparse==0.3.0
vine==1.3.0
zipp==0.6.0
kafka-python==1.4.7
pykafka

